Moved fetchAllStopAreas to stopAreaService
